+++ SDK_C++/CodeCraft-2021/vm_manager.cpp
@@ -12,7 +12,7 @@ VmManager::VmManager() :
   request_info_list_(request_data_manager_.GetRequestInfoList()) {
     vm_schedules_.reserve(num_vm_);
     for (const auto& request : request_info_list_) {
-        const int16_t request_id = request.first;
+        const int32_t request_id = request.first;
         const RequestInfo& request_info = request.second;
         int vm_id = vm_data_manager_.GetVmId(request_info.requested_vm_name);
