+++ SDK_C++/CodeCraft-2021/vm_manager.h
@@ -19,10 +19,10 @@
 // and the schedules of these vm's
 struct VmStatusInfo {
     std::vector<uint32_t> num_running;
-    std::unordered_map<uint32_t, std::vector<int32_t> > request_id_list; // Day, list of server id
+    std::unordered_map<int32_t, std::vector<int32_t> > request_id_list; // Day, list of server id
 
     // Day, list of server id; currently unused, but might need it later when mapping VM's to servers
-    std::unordered_map<uint32_t, std::vector<int32_t> > server_id_list;
+    std::unordered_map<int32_t, std::vector<int32_t> > server_id_list;
 };
 
 
@@ -41,7 +41,7 @@ class VmManager {
     RequestDataManager& request_data_manager_;
     uint16_t days_;
     uint16_t num_vm_;
-    std::unordered_map<int16_t, RequestInfo>& request_info_list_;
+    std::unordered_map<int32_t, RequestInfo>& request_info_list_;
     // vm id to a list of schedules of all vm's of this kind
     // TODO(Yu Xin): use next_available when adding requests to the list
     std::unordered_map<uint16_t, VmStatusInfo> vm_schedules_;
